backend/requirements.txt
# FastAPI production server
gunicorn==21.2.0
uvicorn[standard]==0.27.0

# Database
sqlmodel==0.0.16
psycopg2-binary==2.9.9

# Async support
asyncpg==0.29.0

# Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.0
python-multipart==0.0.6

# Environment variables
pydantic-settings==2.1.0

# CORS
python-dotenv==1.0.0
